requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.9
aiosignal==1.3.1
async-timeout==5.0.1
attrs==24.2.0
beautifulsoup4==4.12.3
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
pika==1.3.2
pika-stubs==0.1.3
propcache==0.2.1
python-dotenv==1.0.1
soupsieve==2.6
typing_extensions==4.12.2
yarl==1.18.3
